refactor : 완등률 double -> int

+++ clog-api/src/main/kotlin/org/depromeet/clog/server/api/report/application/GetReport.kt
@@ -39,7 +39,7 @@ class GetReport(
         val totalAttemptCount = TotalAttemptCount(
             successAttemptCount = reportQuery.successAttemptCount,
             totalAttemptCount = reportQuery.totalAttemptCount,
-            completionRate = completionRate
+            completionRate = completionRate.toInt()
         )
 
         var statistic = dailyReportStatisticRepository.findByUserId(userId)
